Toolbar cleanup
