+++ src/components/toolbar/index.tsx
@@ -1,14 +1,12 @@
 import { connect } from 'react-redux';
 
-import Renderer from './renderer';
-
 import * as EditorActions from '../../actions/editor';
 import { State } from '../../constants/default-state';
+import Renderer from './renderer';
 
 function mapStateToProps(state: State, ownProps) {
   return {
     error: state.error,
-    export: state.export,
     mode: state.mode,
     renderer: state.renderer,
     warningsLogger: state.warningsLogger,
@@ -17,9 +15,6 @@ function mapStateToProps(state: State, ownProps) {
 
 const mapDispatchToProps = dispatch => {
   return {
-    exportVega: val => {
-      dispatch(EditorActions.exportVega(val));
-    },
     setRenderer: val => {
       dispatch(EditorActions.setRenderer(val));
     },
